# support NHWC format  and improve performance of temporal_shift op